Renumber to DAP053/DAP054: the runtime-bridge PR is closed, so its id is free
The migration diagnostic becomes DAP053 and the unusable-handler refusal
DAP054, keeping the library block contiguous now that nothing is going to take
DAP053 for [UseRuntimeTypeHandlers].
