+++ docs/rules/DAP054.md
@@ -1,41 +1,33 @@
 # DAP054
 
-Type-handler registered at runtime
+Type-handler cannot be used
 
-`SqlMapper.AddTypeHandler` mutates a process-global registry when it executes. The generator
-runs at compile time, so it cannot see that call - and generated code therefore binds the value
-the way it would have without any handler at all: usually a raw `p.Value = ...`, which the
-provider rejects ("No mapping exists from object type ..."), or a read that never reaches your
-`Parse`.
+A `[TypeHandler(typeof(TValue), typeof(THandler))]` registration named a handler that generated
+code cannot use, so the registration is **ignored** - and values of that type bind as if it were
+not there, which is usually the raw `p.Value = ...` the handler existed to replace.
 
-``` csharp
-SqlMapper.AddTypeHandler(new LocalDateHandler()); // DAP054
-```
+The reasons, and what each means:
 
-**The fix**: declare the handler instead, so the generator can bake the dispatch:
-
-``` csharp
-[module: TypeHandler(typeof(LocalDate), typeof(LocalDateHandler))]
-```
-
-The declarative form is per-assembly rather than process-global, is deterministic (no
-startup-ordering races), and shows up in review. An existing handler written against vanilla
-Dapper (`SqlMapper.TypeHandler<T>` / `SqlMapper.ITypeHandler`) can be named as-is: generated
-code adapts it to Dapper.AOT's `IDbValueHandler<T>`.
-
-A handler written for Dapper.AOT directly implements `IDbValueHandler<T>`, usually by
-inheriting `DbValueHandler<T>`:
+| message | what to do |
+| --- | --- |
+| implements neither `IDbValueHandler<T>` nor `SqlMapper.ITypeHandler` | inherit `DbValueHandler<TValue>`, or name a vanilla Dapper handler (generated code adapts it) |
+| handles '`X`', not '`Y`' | the handler and the registered value type disagree - fix whichever is wrong |
+| it has no public parameterless constructor | generated code constructs the handler itself; give it one, or use a handler that needs no state |
+| it is abstract / it is static | name a concrete, instantiable type |
+| it is not accessible to generated code | make it `public` or `internal` (generated code lives in your own assembly) |
+| the handled type is not accessible to generated code | same, for `TValue` |
 
 ``` csharp
-public sealed class LocalDateHandler : DbValueHandler<LocalDate>
+[module: TypeHandler(typeof(LocalDate), typeof(LocalDateHandler))] // DAP054 if the handler is wrong
+
+public sealed class LocalDateHandler : DbValueHandler<LocalDate> // <-- the shape it wants
 {
-    protected override void Configure(DbParameter parameter) => parameter.DbType = DbType.Date;
     protected override void SetValueCore(DbParameter parameter, LocalDate value)
         => parameter.Value = new DateTime(value.Year, value.Month, value.Day);
     protected override LocalDate Parse(object? value) { /* ... */ }
 }
 ```
 
-Note the limit of this diagnostic: it only sees `AddTypeHandler` calls in the compilation being
-built. A registration performed by a referenced library, or through a helper that hides the
-call, cannot be reported - so its absence is not proof that generated code sees every handler.
+This is a warning rather than an error because the code still compiles and runs - it just binds
+without the handler. It is reported precisely because that failure is otherwise silent, which is
+the thing the declarative registration exists to remove.
