--- docs/rules/DAP055.md
@@ -1,33 +1,0 @@
-# DAP055
-
-Type-handler cannot be used
-
-A `[TypeHandler(typeof(TValue), typeof(THandler))]` registration named a handler that generated
-code cannot use, so the registration is **ignored** - and values of that type bind as if it were
-not there, which is usually the raw `p.Value = ...` the handler existed to replace.
-
-The reasons, and what each means:
-
-| message | what to do |
-| --- | --- |
-| implements neither `IDbValueHandler<T>` nor `SqlMapper.ITypeHandler` | inherit `DbValueHandler<TValue>`, or name a vanilla Dapper handler (generated code adapts it) |
-| handles '`X`', not '`Y`' | the handler and the registered value type disagree - fix whichever is wrong |
-| it has no public parameterless constructor | generated code constructs the handler itself; give it one, or use a handler that needs no state |
-| it is abstract / it is static | name a concrete, instantiable type |
-| it is not accessible to generated code | make it `public` or `internal` (generated code lives in your own assembly) |
-| the handled type is not accessible to generated code | same, for `TValue` |
-
-``` csharp
-[module: TypeHandler(typeof(LocalDate), typeof(LocalDateHandler))] // DAP055 if the handler is wrong
-
-public sealed class LocalDateHandler : DbValueHandler<LocalDate> // <-- the shape it wants
-{
-    protected override void SetValueCore(DbParameter parameter, LocalDate value)
-        => parameter.Value = new DateTime(value.Year, value.Month, value.Day);
-    protected override LocalDate Parse(object? value) { /* ... */ }
-}
-```
-
-This is a warning rather than an error because the code still compiles and runs - it just binds
-without the handler. It is reported precisely because that failure is otherwise silent, which is
-the thing the declarative registration exists to remove.
