+++ test/Dapper.AOT.Test/Verifiers/DAP054.cs
@@ -7,75 +7,94 @@ namespace Dapper.AOT.Test.Verifiers;
 
 public class DAP054 : Verifier<DapperAnalyzer>
 {
-    [Fact] // a runtime registration the generator cannot see, in both spellings
-    public Task RuntimeRegistrationIsInvisible() => CSVerifyAsync("""
+    // module attributes must precede everything else, so each case declares its own header
+    private const string Header = """
         using Dapper;
         using System;
         using System.Data;
+        using System.Data.Common;
 
         [module: DapperAot]
+        """;
+
+    private const string Types = """
 
         public struct LocalDate { public int Year; }
+        public struct Money { public decimal Amount; }
+        """;
 
-        public sealed class LocalDateHandler : SqlMapper.TypeHandler<LocalDate>
-        {
-            public override void SetValue(IDbDataParameter parameter, LocalDate value) { }
-            public override LocalDate Parse(object value) => default;
-        }
+    [Fact] // the handler implements neither contract - the commonest way to get this wrong
+    public Task NeitherContract() => CSVerifyAsync(Header + """
+        [module: {|#0:TypeHandler(typeof(LocalDate), typeof(NotAHandler))|}]
+        """ + Types + """
 
-        public static class Startup
-        {
-            public static void Register()
-            {
-                {|#0:SqlMapper.AddTypeHandler(new LocalDateHandler())|};
-                {|#1:SqlMapper.AddTypeHandler(typeof(LocalDate), new LocalDateHandler())|};
-            }
-        }
+        public sealed class NotAHandler { }
         """, DefaultConfig, [
-            Diagnostic(Diagnostics.RuntimeTypeHandlerRegistration).WithLocation(0).WithArguments("LocalDate", "LocalDateHandler"),
-            Diagnostic(Diagnostics.RuntimeTypeHandlerRegistration).WithLocation(1).WithArguments("LocalDate", "LocalDateHandler"),
+            Diagnostic(Diagnostics.UnusableTypeHandler).WithLocation(0)
+                .WithArguments("NotAHandler", "LocalDate", "it implements neither IDbValueHandler<T> nor SqlMapper.ITypeHandler"),
     ]);
 
-    [Fact] // declared via the attribute: the runtime call is redundant, not wrong, so: quiet
-    public Task DeclaredHandlerIsNotReported() => CSVerifyAsync("""
-        using Dapper;
-        using System;
-        using System.Data;
+    [Fact] // a real handler, registered against the wrong value type: name both sides
+    public Task WrongValueType() => CSVerifyAsync(Header + """
+        [module: {|#0:TypeHandler(typeof(Money), typeof(LocalDateHandler))|}]
+        """ + Types + """
 
-        [module: DapperAot]
+        public sealed class LocalDateHandler : DbValueHandler<LocalDate>
+        {
+            protected override void SetValueCore(DbParameter parameter, LocalDate value) { }
+            protected override LocalDate Parse(object? value) => default;
+        }
+        """, DefaultConfig, [
+            Diagnostic(Diagnostics.UnusableTypeHandler).WithLocation(0)
+                .WithArguments("LocalDateHandler", "Money", "it handles 'LocalDate', not 'Money'"),
+    ]);
+
+    [Fact] // generated code has to construct it
+    public Task NoPublicParameterlessConstructor() => CSVerifyAsync(Header + """
+        [module: {|#0:TypeHandler(typeof(LocalDate), typeof(NeedsArgs))|}]
+        """ + Types + """
+
+        public sealed class NeedsArgs : DbValueHandler<LocalDate>
+        {
+            public NeedsArgs(int scale) { }
+            protected override void SetValueCore(DbParameter parameter, LocalDate value) { }
+            protected override LocalDate Parse(object? value) => default;
+        }
+        """, DefaultConfig, [
+            Diagnostic(Diagnostics.UnusableTypeHandler).WithLocation(0)
+                .WithArguments("NeedsArgs", "LocalDate", "it has no public parameterless constructor"),
+    ]);
+
+    [Fact] // abstract cannot be instantiated
+    public Task Abstract() => CSVerifyAsync(Header + """
+        [module: {|#0:TypeHandler(typeof(LocalDate), typeof(AbstractHandler))|}]
+        """ + Types + """
+
+        public abstract class AbstractHandler : DbValueHandler<LocalDate>
+        {
+            protected override LocalDate Parse(object? value) => default;
+        }
+        """, DefaultConfig, [
+            Diagnostic(Diagnostics.UnusableTypeHandler).WithLocation(0)
+                .WithArguments("AbstractHandler", "LocalDate", "it is abstract"),
+    ]);
+
+    [Fact] // both good shapes stay quiet: native, and vanilla-via-the-shim
+    public Task UsableHandlersAreQuiet() => CSVerifyAsync(Header + """
         [module: TypeHandler(typeof(LocalDate), typeof(LocalDateHandler))]
+        [module: TypeHandler(typeof(Money), typeof(MoneyHandler))]
+        """ + Types + """
 
-        public struct LocalDate { public int Year; }
-
-        public sealed class LocalDateHandler : SqlMapper.TypeHandler<LocalDate>
+        public sealed class LocalDateHandler : DbValueHandler<LocalDate>
         {
-            public override void SetValue(IDbDataParameter parameter, LocalDate value) { }
-            public override LocalDate Parse(object value) => default;
+            protected override void SetValueCore(DbParameter parameter, LocalDate value) { }
+            protected override LocalDate Parse(object? value) => default;
         }
 
-        public static class Startup
+        public sealed class MoneyHandler : SqlMapper.TypeHandler<Money>
         {
-            public static void Register() => SqlMapper.AddTypeHandler(new LocalDateHandler());
-        }
-        """, DefaultConfig, []);
-
-    [Fact] // no Dapper.AOT in play: vanilla-only code is behaving correctly
-    public Task NotReportedWithoutDapperAot() => CSVerifyAsync("""
-        using Dapper;
-        using System;
-        using System.Data;
-
-        public struct LocalDate { public int Year; }
-
-        public sealed class LocalDateHandler : SqlMapper.TypeHandler<LocalDate>
-        {
-            public override void SetValue(IDbDataParameter parameter, LocalDate value) { }
-            public override LocalDate Parse(object value) => default;
-        }
-
-        public static class Startup
-        {
-            public static void Register() => SqlMapper.AddTypeHandler(new LocalDateHandler());
+            public override void SetValue(IDbDataParameter parameter, Money value) { }
+            public override Money Parse(object value) => default;
         }
         """, DefaultConfig, []);
 }
